repos/packages/design-system/src/antigravity_design_system/pinterest_client.py
import os
import logging
from typing import List, Dict, Any, Optional
try:
    from pinterest.pinterestsdk import PinterestSDKClient
    from pinterest.ads.ad_accounts import AdAccount
    PINTEREST_SDK_AVAILABLE = True
except ImportError:
    PINTEREST_SDK_AVAILABLE = False

logger = logging.getLogger(__name__)

class PinterestClient:
    """
    Client for interacting with Pinterest API v5.
    """
    def __init__(
        self,
        app_id: Optional[str] = None,
        app_secret: Optional[str] = None,
        access_token: Optional[str] = None,
        refresh_token: Optional[str] = None
    ):
        self.app_id = app_id or os.getenv("PINTEREST_APP_ID")
        self.app_secret = app_secret or os.getenv("PINTEREST_APP_SECRET")
        self.access_token = access_token or os.getenv("PINTEREST_ACCESS_TOKEN")
        self.refresh_token = refresh_token or os.getenv("PINTEREST_REFRESH_TOKEN")
        
        self.client = None
        if self.access_token and PINTEREST_SDK_AVAILABLE:
            try:
                self.client = PinterestSDKClient.create_client_with_token(self.access_token)
            except Exception as e:
                logger.error(f"Failed to initialize Pinterest SDK client: {e}")
        elif self.access_token and not PINTEREST_SDK_AVAILABLE:
            logger.warning("Pinterest SDK library not installed. Falling back to mock inspiration.")

    def get_inspiration_tokens(self, board_id: Optional[str] = None) -> List[Dict[str, Any]]:
        """
        Fetch pins from a board and extract inspiration tokens (colors, styles).
        """
        if not self.client:
            logger.warning("Pinterest client not initialized. Returning mock inspiration.")
            return self._get_mock_inspiration()
        
        try:
            # Here we would use the SDK to fetch pins
            # For now, we simulate the logic
            pins = [] # self.client.pins.get_pins_from_board(board_id)
            
            # TODO: Implement real extraction logic once SDK is fully functional in env
            return self._get_mock_inspiration()
        except Exception as e:
            logger.error(f"Error fetching inspiration from Pinterest: {e}")
            return self._get_mock_inspiration()

    def _get_mock_inspiration(self) -> List[Dict[str, Any]]:
        return [
            {"type": "color_palette", "colors": ["#E60023", "#BD081C", "#FFFFFF"]},
            {"type": "layout", "style": "masonry-grid"},
            {"type": "typography", "fonts": ["Helvetica Neue", "Arial"]}
        ]
